Store key using Social Sharing
